add backend for map events

+++ app/map-events/mutations/deleteMapEvent.ts
@@ -6,9 +6,11 @@ const DeleteMapEvent = z.object({
   id: z.number(),
 })
 
-export default resolver.pipe(resolver.zod(DeleteMapEvent), resolver.authorize(), async ({ id }) => {
-  // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-  const mapEvent = await db.mapEvent.deleteMany({ where: { id } })
-
-  return mapEvent
-})
+export default resolver.pipe(
+  resolver.zod(DeleteMapEvent),
+  resolver.authorize(),
+  async ({ id }, { session }) => {
+    const mapEvent = await db.mapEvent.deleteMany({ where: { id, userId: session.userId } })
+    return mapEvent
+  }
+)
